# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/crm/contract/addFile": {
            "post": {
                "description": "添加一个新文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRMcontract相关接口"
                ],
                "summary": "添加文件",
                "parameters": [
                    {
                        "description": "data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ContractAddFile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R"
                        }
                    }
                }
            }
        },
        "/crm/contract/create": {
            "post": {
                "description": "添加一个新合同",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRMcontract相关接口"
                ],
                "summary": "添加新合同",
                "parameters": [
                    {
                        "description": "data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Contract"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R"
                        }
                    }
                }
            }
        },
        "/crm/contract/delete": {
            "post": {
                "description": "删除合同信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRMcontract相关接口"
                ],
                "summary": "删除合同信息",
                "parameters": [
                    {
                        "description": "data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Contract"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R"
                        }
                    }
                }
            }
        },
        "/crm/contract/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRMcontract相关接口"
                ],
                "summary": "获取客户列表",
                "parameters": [
                    {
                        "description": "data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DtoContractList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R"
                        }
                    }
                }
            }
        },
        "/crm/contract/update": {
            "post": {
                "description": "修改合同信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRMcontract相关接口"
                ],
                "summary": "修改合同信息",
                "parameters": [
                    {
                        "description": "data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Contract"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R"
                        }
                    }
                }
            }
        },
...
        "/crm/customer/create": {
            "post": {
                "description": "添加一个新用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRM-customer相关接口"
                ],
                "summary": "添加新用户",
                "parameters": [
                    {
                        "description": "data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R"
                        }
                    }
                }
            }
        },
        "/crm/customer/delete": {
            "post": {
                "description": "删除用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRM-customer相关接口"
                ],
                "summary": "删除用户信息",
                "parameters": [
                    {
                        "description": "data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R"
                        }
                    }
                }
            }
        },
        "/crm/customer/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRM-customer相关接口"
                ],
                "summary": "获取客户列表",
                "parameters": [
                    {
                        "description": "data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DtoCustomerList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R"
                        }
                    }
                }
            }
        },
        "/crm/customer/update": {
            "post": {
                "description": "修改用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRM-customer相关接口"
                ],
                "summary": "修改用户信息",
                "parameters": [
                    {
                        "description": "data",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Contract": {
            "type": "object",
            "properties": {
                "customerId": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "entity.ContractAddFile": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "entity.Customer": {
            "type": "object",
            "properties": {
                "communication": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "entity.DtoContractList": {
            "type": "object",
            "required": [
                "token",
                "version"
            ],
            "properties": {
                "keyword": {
                    "description": "搜索的关键字",
                    "type": "string"
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "size": {
                    "description": "每页显示多少条",
                    "type": "integer"
                },
                "token": {
                    "type": "string",
                    "example": "xxxx"
                },
                "version": {
                    "type": "string",
                    "example": "1001"
                }
            }
        },
        "entity.DtoCustomerList": {
            "type": "object",
            "required": [
                "token",
                "version"
            ],
            "properties": {
                "keyword": {
                    "description": "搜索的关键字",
                    "type": "string"
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "size": {
                    "description": "每页显示多少条",
                    "type": "integer"
                },
                "token": {
                    "type": "string",
                    "example": "xxxx"
                },
                "version": {
                    "type": "string",
                    "example": "1001"
                }
            }
        },
...
        "response.R": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}